Time: 0 ms (100.00%), Space: 6.9 MB (91.95%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.cpp
@@ -1,12 +1,12 @@
 class Solution {
 public:
     bool judgeSquareSum(int c) {
-        unsigned long long left = 0, right = sqrt(c);
-        while(left <= right) {
-            unsigned long long sqsm = left * left + right * right;
-            if(sqsm == c) {
+        long left = 0, right = static_cast<long>(sqrt(c));
+        while (left <= right) {
+            long long sqsm = static_cast<long long>(left) * left + static_cast<long long>(right) * right;
+            if (sqsm == c) {
                 return true;
-            } else if(sqsm > c) {
+            } else if (sqsm > c) {
                 right--;
             } else {
                 left++;
